MultiBranch docs and tidy up.
Move getScriptPathOrDefault and getWorkspace from Helix Branches to
AbstractP4ScmSource.
